- 天气界面背景调整 - 按位置进行搜索 - 位置显示

+++ src/main/java/in/hocg/web/modules/weather/body/Weather.java
@@ -168,8 +168,8 @@ public class Weather {
          */
         
         private double temp;
-        private int pressure;
-        private int humidity;
+        private double pressure;
+        private double humidity;
         private double temp_min;
         private double temp_max;
         
@@ -181,19 +181,19 @@ public class Weather {
             this.temp = temp;
         }
         
-        public int getPressure() {
+        public double getPressure() {
             return pressure;
         }
         
-        public void setPressure(int pressure) {
+        public void setPressure(double pressure) {
             this.pressure = pressure;
         }
         
-        public int getHumidity() {
+        public double getHumidity() {
             return humidity;
         }
         
-        public void setHumidity(int humidity) {
+        public void setHumidity(double humidity) {
             this.humidity = humidity;
         }
         
@@ -220,22 +220,22 @@ public class Weather {
          * deg : 10
          */
         
-        private int speed;
-        private int deg;
+        private double speed;
+        private double deg;
         
-        public int getSpeed() {
+        public double getSpeed() {
             return speed;
         }
         
-        public void setSpeed(int speed) {
+        public void setSpeed(double speed) {
             this.speed = speed;
         }
         
-        public int getDeg() {
+        public double getDeg() {
             return deg;
         }
         
-        public void setDeg(int deg) {
+        public void setDeg(double deg) {
             this.deg = deg;
         }
     }
